requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
joblib>=1.3.0
schedule>=1.2.0
yagmail>=0.15.0
pyyaml>=6.0.1
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
ipykernel>=6.25.0
python-dotenv>=1.0.0
streamlit>=1.35.0
statsmodels>=0.14.0
